requirements.txt
requests
flask
python-dotenv
google-generativeai
streamlit
